tool_packages/unique_web_search/pyproject.toml
[project]
name = "unique_web_search"
version = "1.17.0"
description = ""
readme = "README.md"
requires-python = ">=3.12,<4.0"
authors = [
    { name = "Andreas Hauri", email = "andreas@unique.ch" },
    { name = "Gustav Hartz", email = "gustav.hartz.ext@unique.ch" },
    { name = "Rami Azouz", email = "rami.ext@unique.ch" },
]
license = { text = "Proprietary" }
dependencies = [
    "typing-extensions>=4.14.1,<5",
    "pydantic>=2.12.3,<3",
    "pydantic-settings>=2.10.1,<3",
    "timeout-decorator>=0.5.0,<1",
    "markdownify>=0.14.1,<1",
    "fake-useragent>=2.2.0,<3",
    "crawl4ai>=0.6.3,<1",
    "firecrawl>=3.3.2,<4",
    "tavily-python>=0.7.11,<1",
    "unidecode>=1.4.0,<2",
    "azure-ai-projects>=1.0.0,<2",
    "azure-identity>=1.25.0,<2",
    "unique-toolkit[monitoring]>=1.69.6,<2",
    "azure-core>=1.36.0,<2",
    "google-cloud-aiplatform>=1.128.0,<2",
    "google-auth>=2.43.0,<3",
    "google-generativeai>=0.8.5,<0.9",
    "rapidfuzz>=3.0.0,<4",
    "httpx>=0.28.1",
    "certifi>=2025.11.12",
    "jinja2>=3.1.6",
    "click>=8.1.0,<9",
]

[project.scripts]
unique-websearch = "unique_web_search.cli.cli:main"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = []

[tool.uv]
exclude-newer = "2 weeks"

[tool.uv.exclude-newer-package]
"unique-toolkit" = false

[tool.poe.tasks]
lint = "ruff check ."
lint-fix = "ruff check . --fix"
format = "ruff format ."
test = "pytest"
typecheck = "basedpyright"
depcheck = "deptry ."
coverage = "pytest --cov=unique_web_search --cov-report=term-missing"
ci-typecheck = { shell = "bash $(git rev-parse --show-toplevel)/.github/scripts/dev.sh typecheck" }
ci-coverage = { shell = "bash $(git rev-parse --show-toplevel)/.github/scripts/dev.sh coverage" }

[tool.deptry]
known_first_party = ["unique_web_search"]

[tool.basedpyright]
typeCheckingMode = "standard"
include = ["unique_web_search"]
exclude = ["**/tests/**", "**/test_*.py"]

[tool.pytest.ini_options]
addopts = "--strict-markers --import-mode=importlib"
asyncio_mode = "auto"
markers = [
    "ai: AI-authored or AI-generated tests",
    "asyncio: asyncio tests",
    "integration: integration tests that require API access or credentials",
    "serial: tests that must run serially",
    "unit: unit tests",
    "verified: AI-generated tests with human verification",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
extend-select = ["I"]

[tool.uv.sources]
unique-toolkit = { workspace = true }
